# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for damarmunda, a camunda golang",
        "title": "damarmunda API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/v1",
    "paths": {
...
        "/camunda/incident/{incidentKey}/resolve": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Resolve Incident Bpmn Camunda",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camunda/incident"
                ],
                "summary": "Resolve Incident Bpmn Camunda",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "incidentKey",
                        "name": "incidentKey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/camunda/process-instance": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Proses Instance from rest api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camunda/process-instance"
                ],
                "summary": "Get Proses Instance from rest api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Size 50",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SearchAfter 1731486859777,2251799814109407",
                        "name": "searchAfter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SearchBefore 1731486859777,2251799814109407",
                        "name": "searchBefore",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "search process instance",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create Proses Instance from rest api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camunda/process-instance"
                ],
                "summary": "Create Proses Instance from rest api",
                "parameters": [
                    {
                        "description": "Create Proses Instance Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateProcessInstancePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.CreateProcessInstancesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/camunda/process-instance/{processinstanceKey}/cancel": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cancel Proses Instance from rest api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camunda/process-instance"
                ],
                "summary": "Cancel Proses Instance from rest api",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "processinstanceskey",
                        "name": "processinstanceKey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/camunda/resource/crud": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "CRUD Store And Handler from Payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camunda/resource"
                ],
                "summary": "CRUD Store And Handler",
                "parameters": [
                    {
                        "description": "Crud Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CrudPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/camunda/resource/deploy": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deploy Only Bpmn Camunda by Name In Folder Resources",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camunda/resource"
                ],
                "summary": "Deploy Only Bpmn Camunda",
                "parameters": [
                    {
                        "description": "Deploy Bpmn Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.DeployBpmnPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/camunda/resource/deploy-crud": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deploy Bpmn Camunda by Name In Folder Resources And Create CRUD in Store And Handler File",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camunda/resource"
                ],
                "summary": "Deploy Bpmn Camunda and Create CRUD in Store And Handler File",
                "parameters": [
                    {
                        "description": "Deploy Bpmn Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.DeployBpmnPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/camunda/resource/operate/statistics": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Information operate statistics camunda from rest api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camunda/process-instance"
                ],
                "summary": "Get Information operate statistics camunda from rest api",
                "responses": {
                    "200": {
                        "description": "search process instance",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/camunda/resource/{processDefinitionKey}/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete Bpmn Camunda",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camunda/resource"
                ],
                "summary": "Delete Bpmn Camunda",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Process Definition Key",
                        "name": "processDefinitionKey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/camunda/user-task": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Search TaskList from rest api v1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "camunda/user-task"
                ],
                "summary": "Search TaskList from rest api v1",
                "parameters": [
                    {
                        "description": "Search TaskList Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SearchTaskListPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SearchTaskListPayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.CreateProcessInstancePayload": {
            "type": "object",
            "required": [
                "processDefinitionKey"
            ],
            "properties": {
                "awaitCompletion": {
                    "type": "boolean"
                },
                "fetchVariables": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "operationReference": {
                    "type": "integer"
                },
                "processDefinitionKey": {
                    "type": "integer"
                },
                "requestTimeout": {
                    "type": "integer"
                },
                "startInstructions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.StartInstruction"
                    }
                },
                "tenantId": {
                    "type": "string"
                },
                "variables": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "main.StartInstruction": {
            "type": "object",
            "properties": {
                "elementId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}